Create separate file for MusicAlbum and Genre's methods
